+++ assets/javascripts/kitten/components/form/text-input-with-unit/index.js
@@ -131,7 +131,7 @@ export const TextInputWithUnit = ({
           },
         )}
       >
-        {unit || unitWord }
+        {unit || unitWord}
       </span>
     </StyledTextInputWithUnit>
   )