+++ assets/javascripts/kitten/karl/pages/recurring-donation-card/stories.js
@@ -1,5 +1,11 @@
 import React from 'react'
-import { BackingCard, TextInputWithUnit, Text, Title, Label } from 'kitten'
+import {
+  BackingCard,
+  TextInputWithUnit,
+  Text,
+  Title,
+  Label,
+} from 'kitten'
 import { StyledRecurringDonation } from './styles'
 
 export default {
@@ -8,13 +14,7 @@ export default {
 
 export const DonRecurent = (props) => {
   return (
-    <div style={{
-      display: 'grid',
-      alignItems: 'stretch',
-      gridTemplateColumns: '1fr 1fr 1fr',
-      gridAutoRows: '300px',
-      gap: '10px',
-    }}>
+    <div className="story-Container story-Grid">
       <StyledRecurringDonation>
         <BackingCard {...props}>
           <BackingCard.Title>
@@ -59,42 +59,6 @@ export const DonRecurent = (props) => {
           </BackingCard.Button>
         </BackingCard>
       </StyledRecurringDonation>
-
-      <StyledRecurringDonation>
-        <BackingCard {...props}>
-          <BackingCard.Title>
-            <Title tag="h3" modifier="quaternary" noMargin>
-              Don mensuel
-            </Title>
-          </BackingCard.Title>
-
-          <BackingCard.Form>
-            <Text tag="p" size="default" className="k-u-margin-bottom-double k-u-margin-top-none">
-              Faites un don mensuel du montant de votre choix pour continuer à faire vivre notre projet&nbsp;:
-            </Text>
-            <div className="k-RecurringDonation__thumbnailsTrio">
-              <div as="button" className="k-RecurringDonation__thumbnail">
-                <Text tag="p" className="k-u-margin-none" weight="bold" size="huge">
-                  2&nbsp;€
-                </Text>
-                <Text className="k-u-margin-none" tag="p" size="tiny">par mois</Text>
-              </div>
-              <div as="button" className="k-RecurringDonation__thumbnail">
-                <Text className="k-u-margin-none" tag="p" weight="bold" size="huge">5&nbsp;€</Text>
-                <Text className="k-u-margin-none" tag="p" size="tiny">par mois</Text>
-              </div>
-              <div as="button" className="k-RecurringDonation__thumbnail">
-                <Text tag="p" className="k-u-margin-none" weight="bold" size="huge">10&nbsp;€</Text>
-                <Text className="k-u-margin-none" tag="p" size="tiny">par mois</Text>
-              </div>
-            </div>
-          </BackingCard.Form>
-  
-          <BackingCard.Button className="k-u-margin-bottom-triple">
-            Nous soutenir
-          </BackingCard.Button>
-        </BackingCard>
-      </StyledRecurringDonation>
-    </div> 
+    </div>
   )
 }